+++ settings.h
@@ -9,15 +9,15 @@ struct node {
 extern GSettings *settings;
 
 void
-settings_save_nodes(struct node *nodes, int len);
+settings_save_nodes(struct node *nodes, size_t len);
 
 int
 settings_load_nodes();
 
 void
-settings_save_node_id(char *id, int len);
+settings_save_node_id(char *id, size_t len);
 
 void
-settings_load_node_id(const char *id, unsigned long len);
+settings_load_node_id(const char *id, size_t len);
 
 #endif
